ステートとアクションのクリエイター作成、カメラの揺れ追加、回避アクション途中

+++ CW3D/Project/CameraController.h
@@ -73,7 +73,7 @@
 			return m_Camera->GetCamera();
 		}
 
-		void Quake(float power, float freq, int time);
+		void Quake(float power, float freq, float time);
 		void Enable(const Vector3& pos, const Vector3& lookPos);
 
 	};
